+++ src/opts/pods.rs
@@ -309,6 +309,11 @@ impl PodCreateOptsBuilder {
     );
 
     impl_vec_field!(
+        /// The ID of the pod's service container.
+        service_container_id => "serviceContainerID"
+    );
+
+    impl_vec_field!(
         /// Instructs the pod to share a set of namespaces. Shared namespaces will be joined (by
         /// default) by every container which joins the pod. If not set and NoInfra is false, the
         /// pod will set a default set of namespaces to share. Conflicts with